+++ src/main/kotlin/stacrypt/stawallet/ripple/RippleWallet.kt
@@ -3,18 +3,13 @@ package stacrypt.stawallet.ripple
 import com.typesafe.config.Config
 import org.jetbrains.exposed.sql.select
 import org.jetbrains.exposed.sql.transactions.transaction
-import org.kethereum.encodings.encodeToBase58WithChecksum
-import org.kethereum.extensions.toBytesPadded
-import org.kethereum.extensions.toMinimalByteArray
 import stacrypt.stawallet.ConfigSecretProvider
 import stacrypt.stawallet.NotEnoughFundException
 import stacrypt.stawallet.Wallet
-import stacrypt.stawallet.bitcoin.toBitcoinWif
 import stacrypt.stawallet.model.AddressDao
 import stacrypt.stawallet.model.AddressTable
 import stacrypt.stawallet.model.DepositDao
 import stacrypt.stawallet.model.InvoiceDao
-import java.math.BigInteger
 import kotlin.math.roundToLong
 
 const val LEDGER_INDEX_VALIDATED = "validated"
@@ -59,7 +54,7 @@ class RippleWallet(name: String, config: Config, network: String) : Wallet(
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override suspend fun sendTo(address: String, amountToSend: BigInteger, tag: Any?): Any = transaction {
+    override suspend fun sendTo(address: String, amountToSend: Int, tag: Any?): Any = transaction {
         val accountInfo = daemon.rpcClient.getAccountInfo(
             account = theOnlyHotAddress!!.provision,
             ledgerIndex = LEDGER_INDEX_VALIDATED
